packages/sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "comfyui-agent-sdk"
version = "0.1.0"
description = "Python SDK for ComfyUI workflow management, model management, and automation"
readme = "../../README.md"
requires-python = ">=3.10"
license = "Apache-2.0"

dependencies = [
    "requests>=2.31.0",
    "mcp>=1.0.0",
    "huggingface-hub>=0.20.0",
    "keyring>=25.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
websocket = ["websocket-client>=1.6.0"]
recommender = ["ollama>=0.1.0"]
dev = ["pytest>=7.4.0", "ruff>=0.1.0", "mypy>=1.7.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/comfyui_agent_sdk"]
